# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tidarr API",
    "description": "Tidarr REST API for managing Tidal music downloads, queue management, synchronization, and Lidarr integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8484",
      "description": "Local Tidarr instance"
    }
  ],
  "security": [
    {
      "ApiKeyHeader": []
    },
    {
      "ApiKeyQuery": []
    },
    {
      "BearerToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key passed via X-Api-Key header (recommended for automation)"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "API key passed via query parameter"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth"
      }
    },
    "schemas": {
      "ContentType": {
        "type": "string",
        "enum": [
          "album",
          "track",
          "playlist",
          "artist",
          "video",
          "favorite_albums",
          "favorite_tracks",
          "favorite_playlists",
          "favorite_videos",
          "favorite_artists",
          "artist_videos",
          "mix"
        ],
        "description": "Type of Tidal content"
      },
      "QualityType": {
        "type": "string",
        "enum": [
          "low",
          "normal",
          "high",
          "max"
        ],
        "description": "Download quality. low=AAC-96, normal=AAC-320, high=16-bit FLAC, max=24-bit FLAC"
      },
      "ProcessingItem": {
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "type",
          "quality",
          "status",
          "loading",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Tidal content ID (numeric string for albums/tracks/artists/videos, UUID for playlists)"
          },
          "title": {
            "type": "string",
            "description": "Item title. Use 'All albums' for artist type, 'All artist videos' for artist_videos type"
          },
          "artist": {
            "type": "string",
            "description": "Artist name. Can be empty string for playlists, mixes, and favorites"
          },
          "url": {
            "type": "string",
            "description": "Full Tidal URL (e.g. https://listen.tidal.com/album/123456) or content path. Optional for favorite_* types"
          },
          "type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityType"
          },
          "status": {
            "type": "string",
            "enum": [
              "queue_download",
              "download",
              "queue_processing",
              "processing",
              "queue",
              "finished",
              "error"
            ],
            "description": "Item status. Use 'queue_download' when adding new items. 'queue' is legacy"
          },
          "loading": {
            "type": "boolean",
            "description": "Whether the item is loading. Set to true when adding new items"
          },
          "error": {
            "type": "boolean",
            "description": "Whether the item has an error. Set to false when adding new items"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/save": {
      "post": {
        "operationId": "addToQueue",
        "summary": "Add an item to the download queue",
        "description": "Add an album, track, video, playlist, mix, artist discography, or favorites to the download queue. The item object must include all required fields: id, title, artist, url, type, quality, status, loading, and error.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/ProcessingItem"
                  }
                }
              },
              "examples": {
                "album": {
                  "summary": "Download an album",
                  "value": {
                    "item": {
                      "id": "251082404",
                      "title": "Random Access Memories",
                      "artist": "Daft Punk",
                      "url": "https://listen.tidal.com/album/251082404",
                      "type": "album",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                },
                "track": {
                  "summary": "Download a track",
                  "value": {
                    "item": {
                      "id": "123456789",
                      "title": "Get Lucky",
                      "artist": "Daft Punk",
                      "url": "https://listen.tidal.com/track/123456789",
                      "type": "track",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                },
                "video": {
                  "summary": "Download a video",
                  "value": {
                    "item": {
                      "id": "123456789",
                      "title": "Video Title",
                      "artist": "Artist Name",
                      "url": "123456789",
                      "type": "video",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                },
                "playlist": {
                  "summary": "Download a playlist",
                  "value": {
                    "item": {
                      "id": "abc123-def456",
                      "title": "My Playlist",
                      "artist": "",
                      "url": "https://listen.tidal.com/playlist/abc123-def456",
                      "type": "playlist",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                },
                "artist": {
                  "summary": "Download all albums by an artist",
                  "value": {
                    "item": {
                      "id": "3566315",
                      "title": "All albums",
                      "artist": "Daft Punk",
                      "url": "https://listen.tidal.com/artist/3566315",
                      "type": "artist",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                },
                "favorite_albums": {
                  "summary": "Download all favorite albums (url is optional for favorites)",
                  "value": {
                    "item": {
                      "id": "favorite_albums",
                      "title": "Favorite Albums",
                      "artist": "",
                      "type": "favorite_albums",
                      "quality": "high",
                      "status": "queue_download",
                      "loading": true,
                      "error": false
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to queue"
          },
          "400": {
            "description": "Invalid request body or item structure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/remove": {
      "delete": {
        "operationId": "removeFromQueue",
        "summary": "Remove an item from the queue",
        "description": "Remove a specific item from the download queue by its ID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "description": "ID of the item to remove"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Item removed"
          },
          "400": {
            "description": "Invalid or missing ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/single-download": {
      "post": {
        "operationId": "singleDownload",
        "summary": "Trigger a single download in NO_DOWNLOAD mode",
        "description": "Trigger a one-off download for a specific item. Works for items in 'queue_download', 'error', or 'finished' status. The item goes through the full download pipeline (beets, move to library, notifications) and ends up as 'finished'. Useful in NO_DOWNLOAD mode to selectively download individual items.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "description": "ID of the item to download"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Download triggered"
          },
          "400": {
            "description": "Invalid or missing ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Item not found or not in valid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}